Move SelfDestructingMessageView to support package.
