+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/human/HumanActivity.java
@@ -7,7 +7,7 @@ import android.support.v7.app.AppCompatActivity;
 import com.novoda.notils.caster.Views;
 import com.novoda.tpbot.Direction;
 import com.novoda.tpbot.R;
-import com.novoda.tpbot.SelfDestructingMessageView;
+import com.novoda.tpbot.support.SelfDestructingMessageView;
 import com.novoda.tpbot.support.SwitchableView;
 
 import java.util.concurrent.TimeUnit;
