+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/support/SelfDestructingMessageView.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot;
+package com.novoda.tpbot.support;
 
 import android.content.Context;
 import android.util.AttributeSet;
