+++ TelepresenceBot/mobile/src/main/res/layout-land/activity_human.xml
@@ -7,7 +7,7 @@
   android:orientation="horizontal"
   tools:context="com.novoda.tpbot.human.HumanActivity">
 
-  <com.novoda.tpbot.SelfDestructingMessageView
+  <com.novoda.tpbot.support.SelfDestructingMessageView
     android:id="@+id/bot_controller_debug_view"
     style="@style/Controller.DebugText"
     android:layout_width="0dp"
